requirements.txt
requests
pandas
psycopg2-binary
python-dotenv
apscheduler